+++ src/OpenXR.hpp
@@ -7,24 +7,24 @@
 #include <openxr_platform.h>
 #include <openxr_reflection.h>
 
+struct OpenXRRenderContext {
+    XrSwapchain swapchains[2];
+    std::vector<XrSwapchainImageVulkanKHR> swapchainImages[2];
+    std::array<XrView, 2> views;
+    std::array<XrCompositionLayerProjectionView, 2> projectionViews;
+};
+
 class OpenXR : public VRInterface {
 private:
     XrSession session;
     XrInstance instance;
     XrSystemId systemId;
-    XrSwapchain swapchains[2];
     XrSpace space, viewSpace;
     XrFrameState frameState;
-    std::vector<XrSwapchainImageVulkanKHR> swapchainImages[2];
-    std::array<XrView, 2> views;
-    std::array<XrCompositionLayerProjectionView, 2> projectionViews;
     int64_t swapchainFormat;
     XrPosef viewPose;
     bool hasProjection;
     bool resetViewRequested;
-
-    uint32_t renderWidth[2];
-    uint32_t renderHeight[2];
 
     std::vector<char> deviceExtensions;
     std::vector<char> instanceExtensions;
@@ -40,6 +40,10 @@ private:
     void UpdatePoses();
     bool GetProjectionMatrix();
     void RecenterView();
+    OpenXRRenderContext* XrContext()
+    {
+        return reinterpret_cast<OpenXRRenderContext*>(currentRenderContext->ext);
+    }
 
 public:
     OpenXR();
@@ -58,10 +62,6 @@ public:
     bool UpdateVRPoses(Quaternion* carQuat, Config::HorizonLock lockSetting) override;
     void PrepareFramesForHMD(IDirect3DDevice9* dev) override;
     void SubmitFramesToHMD(IDirect3DDevice9* dev) override;
-    std::tuple<uint32_t, uint32_t> GetRenderResolution(RenderTarget tgt) const override
-    {
-        return std::make_tuple(renderWidth[0], renderHeight[0]);
-    }
     void ResetView() override;
     FrameTimingInfo GetFrameTiming() override;
     VRRuntime GetRuntimeType() const override { return OPENXR; }